Fix bug where failing consumed characters
If parser fails, we don't want to do anything - the parser should just
return the original input.

Bug was in "sequence" implementation and had matching error in test
assertion.

Refactored tests so that fail_input is explicit and test checks that
"rest" == fail_input to prevent me from making mistakes in writing
the string for what fail_test["rest"] should be.