Add ability for parsers to push to a stack
Having low-level parsers push data onto a stack will allow higher-level
parsers to, on successful parsing, pop values from the stack to perform
calculations. For example, the weekday parser pushes the weekday onto
the stack, so  a day_range function could match a day, dash,
and a day, then pop the two days off the stack to determine the range.

Switched order of parsers and pass_input in test_either() to test the
stack logic in either.