Add day_range parser
Decided to have day_range push a list of days in the range onto the
stack. I'm planning to have the time parser return a start time and end
time, so the next level up parser can do:
```
for day in days:
    start_datetime = day + start_time
    end_datetime = day + endtime
```

Having the range of days expanded into a list of individual days at the
day_range parser will mean that the ranges don't need to be expanded
before entering the above for loop.