Fix bug in how n_or_more handles end of string
In previous code, if n_or_more successfully parsed all the way to the
end of the string, it would then try to parse a zero-length string and
raise an "Index Out Of Range" exception. The while loop will now detect
if the string is empty and trigger deciding if the n_or_more block
succeeded.

A test for matching to the end of the string was added.

In addition, the "rest" returned if n_or_more is successful AND
n_or_more parsed to the end of the string should be the "rest" returned
from the last parse, not "next". If n_or_more fails to match, "next" and
"rest" will be equal, but if the last match was successful, "next" will
be one parse behind "rest". Return statement on line 262 was updated
to return "rest" instead of "next".