Add ModularDateTime class with input tests
The goal of this tool is to determine whether a given day and time is
within the open hours of a restaurant, so the current time will need to
be compared with the open and close times of each restaurant.
This could be handled by comparing the day, then hour, then minute to
check that rest_open <= current_datetime < rest_close, but there
are may edge cases for when the hour overflows into the next day, or
when the day overflows into the next week.

I think this problem can be elegantly solved by creating a
ModularDateTime class that encodes the datetime as a modular integer
representing the number of seconds since 00:00 on Monday.

With modular arithmetic, testing if current time is in open hours can
be done by testing
```
if (current_datetime - rest_open) < (rest_close - rest_open)
```
which should work correctly even in situations of overflow.