Move handling of empty strings to lowest parsers
Handling of empty strings was previously attempted from higher-level
parsers such as "sequence" and "either". However, this produces some
weird bugs in edge cases, such as a sequence not running an
n_or_more(n=0) parser on an empty string. It's simpler to have the
lowest-level parsers detect and empty string and return a parse failure.