Refactor output structure from parsers
Output structure was previously:
{
    "success": bool,
    "rest": string,
    "stack" (optional): [{ <items> }]
}

A parser would be called and the result stored with

result = parser(input)

The result would then be checked for if `result["success"] == True`, and
if so, further operations would be performed on the "stack". In
addition, the result["rest"] parameter was trusted to be accurate and
sometimes used in subsequent steps even if the parser failed.

However, this structure is inefficient - the parser could instead just
return None if parsing fails and (data, rest) if parsing succeeds. This
way the next-level-up parser can detect parse success with

if result := parser(input)

and proceed accordingly. The "success" parameter becomes unnecessary,
and having the order of terms be standardized removes the need for using
dictionary keys to identify the different data returned.

All parsers were refactored to return (data, rest) and to expect
(data, rest) to be returned from lower-level parsers. In addition,
the tests for each parser were refactored to reflect the new data
structure.