Fix week overflow bug in day_range
I had decided to covert days to ModularDatetime ranges in the high-level
parser datetime, but this allowed a bug to occur if a date range was
specified that overflowed across the week boundary, e.g. "Fri-Mon".
This is because the day_range parser was naively creating a range from
start_day to end_day, which will return an empty range if end_day <
start_day.

If the weekday parser returns a ModularDatetime object instead of an
int, the correct modular math can be applied for subsequent parsing
operations.

The weekday parser is changed to return days as a list of
ModularDatetime objects. The day_range parser is changed to generate a
range of ModularDatetime objects correctly. The datetime parser is
changed to expect days found to already be ModularDatetime objects.

Tests in test_weekday,  test_sequence, test_n_or_more, test_day_range,
and test_days are updated to expect days as ModularDatetime objects.
A test for week overflow in day_range is also added to catch this bug
in future.